# libs/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Trader API",
    "summary": "Исследовательская платформа по рыночным данным",
    "description": "REST API платформы. Документация: `/docs` (Swagger UI), `/redoc`, схема — `/openapi.json`.\n\n**WebSocket** `/ws/jobs/{job_id}` (в OpenAPI не описывается): шлёт JSON в формате `JobOut` при каждом изменении задачи и закрывается по её завершении; код закрытия 4404 — задачи нет.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/analogues": {
      "get": {
        "tags": [
          "analogues"
        ],
        "summary": "Исторические аналоги формации",
        "description": "Ближайшие по DTW исторические формации к запросу — вхождению паттерна или события уровня (`key`) либо окну последних `window` баров ряда прогона `query_run_id`. Формация сжимается до PIP-точек и нормализуется (ATR или процент); расстояние — DTW с полосой Сакоэ–Чибы. История — вхождения из прогонов `run_id` (состав прогонов задаёт область поиска), только то, что известно на `as_of`; окно запроса и пересекающиеся с ним вхождения исключаются. Ответ: аналоги с исходами, статистика MVP-5 по ним, траектории и перцентили 25/50/75. Результат кэшируется.",
        "operationId": "getAnalogues",
        "parameters": [
          {
            "name": "query_run_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Прогон, чей ряд — запрос",
              "title": "Query Run Id"
            },
            "description": "Прогон, чей ряд — запрос"
          },
          {
            "name": "run_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "integer"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Прогоны истории (по умолчанию — запроса)",
              "title": "Run Id"
            },
            "description": "Прогоны истории (по умолчанию — запроса)"
          },
          {
            "name": "key",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Ключ вхождения-запроса; без него — окно",
              "title": "Key"
            },
            "description": "Ключ вхождения-запроса; без него — окно"
          },
          {
            "name": "window",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 10,
              "default": 60,
              "title": "Window"
            }
          },
          {
            "name": "pip_points",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 30,
              "minimum": 3,
              "default": 9,
              "title": "Pip Points"
            }
          },
          {
            "name": "normalization",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "atr",
                "percent"
              ],
              "type": "string",
              "default": "atr",
              "title": "Normalization"
            }
          },
          {
            "name": "k",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "K"
            }
          },
          {
            "name": "max_distance",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Distance"
            }
          },
          {
            "name": "band",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 1,
              "minimum": 0,
              "default": 0.2,
              "title": "Band"
            }
          },
          {
            "name": "horizon",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "integer"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Горизонты, баров",
              "title": "Horizon"
            },
            "description": "Горизонты, баров"
          },
          {
            "name": "include_candidates",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Candidates"
            }
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          },
          {
            "name": "seed",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Seed"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnaloguesOut"
                }
              }
            }
          },
          "404": {
            "description": "Не найдено"
          },
          "422": {
            "description": "Неверные параметры"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnaloguesOut": {
        "properties": {
          "query": {
            "$ref": "#/components/schemas/FormationOut"
          },
          "occurrence_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurrence Key",
            "description": "Ключ вхождения-запроса, если был"
          },
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of",
            "description": "Момент, которым ограничены история и бары"
          },
          "unit": {
            "type": "string",
            "enum": [
              "atr",
              "pct"
            ],
            "title": "Unit"
          },
          "considered": {
            "type": "integer",
            "title": "Considered",
            "description": "Кандидатов до отбора"
          },
          "matches": {
            "items": {
              "$ref": "#/components/schemas/MatchOut"
            },
            "type": "array",
            "title": "Matches"
          },
          "stats": {
            "$ref": "#/components/schemas/StatsOut",
            "description": "Исходы аналогов, статистика MVP-5"
          },
          "percentiles": {
            "items": {
              "$ref": "#/components/schemas/PercentileOut"
            },
            "type": "array",
            "title": "Percentiles",
            "description": "Перцентили 25/50/75 траекторий по барам после входа"
          },
          "trajectory_count": {
            "type": "integer",
            "title": "Trajectory Count"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "query",
          "occurrence_key",
          "as_of",
          "unit",
          "considered",
          "matches",
          "stats",
          "percentiles",
          "trajectory_count",
          "warnings"
        ],
        "title": "AnaloguesOut"
      },
...
      "BucketOut": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "n_occurrences": {
            "type": "integer",
            "title": "N Occurrences"
          },
          "horizons": {
            "items": {
              "$ref": "#/components/schemas/HorizonStatsOut"
            },
            "type": "array",
            "title": "Horizons"
          }
        },
        "type": "object",
        "required": [
          "key",
          "n_occurrences",
          "horizons"
        ],
        "title": "BucketOut"
      },
...
      "FormationOut": {
        "properties": {
          "series_key": {
            "type": "string",
            "title": "Series Key"
          },
          "start": {
            "type": "string",
            "format": "date-time",
            "title": "Start",
            "description": "Начало первого бара окна"
          },
          "end": {
            "type": "string",
            "format": "date-time",
            "title": "End",
            "description": "Закрытие последнего бара окна (точка входа)"
          },
          "shape": {
            "$ref": "#/components/schemas/ShapeOut"
          }
        },
        "type": "object",
        "required": [
          "series_key",
          "start",
          "end",
          "shape"
        ],
        "title": "FormationOut"
      },
...
      "HorizonOutcomeOut": {
        "properties": {
          "horizon": {
            "type": "integer",
            "title": "Horizon"
          },
          "censored": {
            "type": "boolean",
            "title": "Censored"
          },
          "ret_atr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ret Atr"
          },
          "ret_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ret Pct"
          },
          "mfe_atr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mfe Atr"
          },
          "mfe_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mfe Pct"
          },
          "mae_atr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mae Atr"
          },
          "mae_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mae Pct"
          },
          "first_hit": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "target",
                  "invalidated"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "First Hit"
          },
          "ambiguous_bar": {
            "type": "boolean",
            "title": "Ambiguous Bar"
          },
          "crosses_session_gap": {
            "type": "boolean",
            "title": "Crosses Session Gap"
          },
          "crosses_weekend": {
            "type": "boolean",
            "title": "Crosses Weekend"
          }
        },
        "type": "object",
        "required": [
          "horizon",
          "censored",
          "ret_atr",
          "ret_pct",
          "mfe_atr",
          "mfe_pct",
          "mae_atr",
          "mae_pct",
          "first_hit",
          "ambiguous_bar",
          "crosses_session_gap",
          "crosses_weekend"
        ],
        "title": "HorizonOutcomeOut"
      },
...
      "MatchOut": {
        "properties": {
          "occurrence": {
            "$ref": "#/components/schemas/OccurrenceOut"
          },
          "formation": {
            "$ref": "#/components/schemas/FormationOut"
          },
          "distance": {
            "type": "number",
            "title": "Distance"
          },
          "normalized_distance": {
            "type": "number",
            "title": "Normalized Distance",
            "description": "Накопленное / длина пути"
          },
          "similarity": {
            "type": "number",
            "title": "Similarity",
            "description": "1 / (1 + normalized_distance), (0, 1]"
          },
          "path": {
            "items": {
              "items": {
                "type": "integer"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Path",
            "description": "Выравнивание точек (запрос, аналог)"
          },
          "outcomes": {
            "items": {
              "$ref": "#/components/schemas/HorizonOutcomeOut"
            },
            "type": "array",
            "title": "Outcomes"
          },
          "trajectory": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trajectory",
            "description": "Сдвиг close после входа по барам, в единицах запроса"
          }
        },
        "type": "object",
        "required": [
          "occurrence",
          "formation",
          "distance",
          "normalized_distance",
          "similarity",
          "path",
          "outcomes",
          "trajectory"
        ],
        "title": "MatchOut"
      },
...
      "OccurrenceOut": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "engine": {
            "type": "string",
            "title": "Engine"
          },
          "kind": {
            "type": "string",
            "enum": [
              "pattern",
              "level"
            ],
            "title": "Kind"
          },
          "group": {
            "type": "string",
            "title": "Group"
          },
          "direction": {
            "type": "string",
            "enum": [
              "bullish",
              "bearish"
            ],
            "title": "Direction"
          },
          "entry": {
            "type": "string",
            "enum": [
              "confirmed",
              "candidate",
              "touch",
              "break"
            ],
            "title": "Entry"
          },
          "available_at": {
            "type": "string",
            "format": "date-time",
            "title": "Available At"
          },
          "target": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          },
          "invalidated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invalidated At"
          },
          "final_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final State"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "false_breakout": {
            "type": "boolean",
            "title": "False Breakout"
          },
          "quality": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "key",
          "engine",
          "kind",
          "group",
          "direction",
          "entry",
          "available_at",
          "target",
          "invalidated_at",
          "final_state",
          "reason",
          "false_breakout",
          "quality",
          "meta"
        ],
        "title": "OccurrenceOut"
      },
      "PercentileOut": {
        "properties": {
          "q": {
            "type": "integer",
            "title": "Q"
          },
          "values": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Values"
          }
        },
        "type": "object",
        "required": [
          "q",
          "values"
        ],
        "title": "PercentileOut"
      },
...
      "ShapeOut": {
        "properties": {
          "indices": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Indices",
            "description": "Индексы PIP-точек в окне (от 0)"
          },
          "times": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Times",
            "description": "Время точек в долях ширины, 0…1"
          },
          "values": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Values",
            "description": "Цена от первой точки (ATR или %)"
          }
        },
        "type": "object",
        "required": [
          "indices",
          "times",
          "values"
        ],
        "title": "ShapeOut"
      },
      "StatsOut": {
        "properties": {
          "matched": {
            "type": "integer",
            "title": "Matched",
            "description": "Вхождений после фильтров"
          },
          "unit": {
            "type": "string",
            "enum": [
              "atr",
              "pct"
            ],
            "title": "Unit"
          },
          "buckets": {
            "items": {
              "$ref": "#/components/schemas/BucketOut"
            },
            "type": "array",
            "title": "Buckets"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "skipped": {
            "type": "integer",
            "title": "Skipped",
            "description": "Вхождений без бара входа в данных ряда",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "matched",
          "unit",
          "buckets",
          "warnings"
        ],
        "title": "StatsOut"
      },
...
    }
  },
...
}